move types and cleanup
