+++ src/renderer/views/NewProjectView.vue
@@ -5,8 +5,8 @@
   import { useTabsStore } from '../stores/tabs'
   import Divider from '../components/Divider.vue'
   import { useSettingsStore } from '../stores/settings'
-  import { Tab } from '../types/tab.type'
-  import { History } from '../types/history.type'
+  import { Tab } from '../../types/tab.typeype'
+  import { History } from '../../types/history.typeype'
   import TextInput from '../components/TextInput.vue'
   import SecondaryButton from '../components/SecondaryButton.vue'
   import ProjectTile from '../components/ProjectTile.vue'
